Fix addDbEntry open dialog
My previous solution worked in terms of adding entries to the database,
however, given that the addDbEntryButton did not use the same function
as the MembersComponent to open the DialogComponent, the former was
not notified when the dialog closes and, therefore, did not receive
the necessary data to update UI.

Now, the same function is passed to the addDbEntryButtonComponent.
Moreover, so as not to lose the context of 'this' and, in turn, the
members array, that function is now an arrow function. Lastly, in the
MembersComponent, array concatenation is used a technique to trigger
the onChanges mechanism to update UI.